pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisperbridge"
version = "1.0.0"
description = "Desktop application for quick text translation using OCR and GPT API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "WhisperBridge Team", email = "team@whisperbridge.dev"},
]
maintainers = [
    {name = "WhisperBridge Team", email = "team@whisperbridge.dev"},
]
keywords = ["ocr", "translation", "gpt", "desktop-application", "screen-capture"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Utilities",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "customtkinter>=5.2.0",
    "pynput>=1.7.6",
    "pystray>=0.19.4",
    "pillow>=10.0.0",
    "mss>=9.0.1",
    "easyocr>=1.7.0",
    "openai>=1.3.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "pydantic>=2.4.0",
    "keyring>=24.2.0",
    "appdirs>=1.4.4",
    "loguru>=0.7.2",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.23.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "memory-profiler>=0.61.0",
    "line-profiler>=4.1.0",
    "responses>=0.24.0",
    "faker>=19.3.0",
]
build = ["build", "twine"]
docs = ["sphinx", "sphinx-rtd-theme"]

[project.urls]
Homepage = "https://github.com/yourusername/whisperbridge"
Documentation = "https://whisperbridge.readthedocs.io/"
Repository = "https://github.com/yourusername/whisperbridge"
"Bug Reports" = "https://github.com/yourusername/whisperbridge/issues"
Changelog = "https://github.com/yourusername/whisperbridge/blob/main/CHANGELOG.md"

[project.scripts]
whisperbridge = "whisperbridge.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
whisperbridge = ["assets/*", "config/*", "assets/icons/*", "assets/sounds/*", "assets/themes/*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["whisperbridge"]
known_third_party = ["customtkinter", "PIL", "pynput", "pystray", "mss", "easyocr", "openai", "httpx", "tenacity", "pydantic", "keyring", "appdirs", "loguru", "click"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "customtkinter.*",
    "pynput.*",
    "pystray.*",
    "mss.*",
    "easyocr.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra -q --cov=whisperbridge --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]